Removed concrete type checks for back-office published cache details, so now working from interface. Renamed controller and FE assets to more generic name.
