+++ src/Umbraco.Web.UI.Client/src/views/dashboard/settings/publishedsnapshotcache.html
@@ -1,4 +1,4 @@
-﻿<div id="nuCache" style="position: relative;" ng-controller="Umbraco.Dashboard.NuCacheController as vm">
+﻿<div id="nuCache" style="position: relative;" ng-controller="Umbraco.Dashboard.PublishedSnapshotCacheController as vm">
 
     <div ng-show="vm.loading || vm.working" style="background: rgba(255, 255, 255, 0.60); position: absolute; left: 0; right: 0; top: 0; bottom: 0;">
         <umb-load-indicator></umb-load-indicator>
