+++ src/Umbraco.Web.UI.Client/src/views/dashboard/settings/publishedstatus.controller.js
@@ -7,8 +7,6 @@
         $http.get(umbRequestHelper.getApiUrl('publishedStatusBaseUrl', 'GetPublishedStatusUrl')),
         'Failed to get published status url')
     .then(function (result) {
-
-        //result = 'views/dashboard/developer/nucache.html'
         vm.includeUrl = result;
     });
